Refactor search hovering to give information
